Bridge meeple overlay
When playing Fairy/Dragon in combination with bridges, it is possible to
get more than one meeple on the same tile.

For a straight road (WE) with a bridge (NS): when a meeple is placed on
the bridge, and one is placed on the farm (N), only the bridge meeple is
visible. (+ rotations).