Rever Java-Websocket library to snapshot version.
Using same version as used in JCZ 3.x due to networking errors.
Recent Java-Websocket library is unstable and causing message lost.
Which leads to crashed because clients are out of sync.
Problem is that connection is not interrupted and behave like nothing
happens.

It would be nice to implement some mechanism to detect dropped
messages (without losing connection) but old version seems to be
better in connection detection failures anyway.

It's possbile that Java-Websocket will be replaced in fureure by
different library (webbit or Tyrus?)